refactor(speech): share the velay host derivation via service-contracts
Devin review: the platform→velay naming convention existed twice (gateway
speech relay + web live-voice client) and could drift silently. Move
velayHostForPlatformHost to @vellumai/service-contracts/ingress — the
package both build units already depend on — and make the gateway's
velayOriginForPlatformHost a thin scheme-adding wrapper.
