+++ clients/web/src/domains/chat/voice/live-voice/connection.test.ts
@@ -31,7 +31,6 @@ import {
   LiveVoiceTokenError,
   mintLiveVoiceToken,
   resolveLiveVoiceWsUrl,
-  velayHostForPlatformHost,
 } from "./connection";
 
 // ---------------------------------------------------------------------------
@@ -172,35 +171,6 @@ describe("buildLiveVoiceWsUrl", () => {
       buildLiveVoiceWsUrl({ assistantId: "assistant-1", token: "tok-abc" }),
     );
     expect(url.host).toBe("velay.vellum.ai");
-  });
-});
-
-// ---------------------------------------------------------------------------
-// velayHostForPlatformHost — mirrors the gateway's velayOriginForPlatformHost
-// ---------------------------------------------------------------------------
-
-describe("velayHostForPlatformHost", () => {
-  test("maps the prod platform host to prod velay", () => {
-    expect(velayHostForPlatformHost("platform.vellum.ai")).toBe(
-      "velay.vellum.ai",
-    );
-  });
-
-  test("maps env-prefixed platform hosts to their env velay", () => {
-    expect(velayHostForPlatformHost("staging-platform.vellum.ai")).toBe(
-      "velay-staging.vellum.ai",
-    );
-    expect(velayHostForPlatformHost("dev-platform.vellum.ai")).toBe(
-      "velay-dev.vellum.ai",
-    );
-  });
-
-  test("returns null for hosts outside the deployment convention", () => {
-    expect(velayHostForPlatformHost("localhost")).toBeNull();
-    expect(velayHostForPlatformHost("platform.example.com")).toBeNull();
-    expect(
-      velayHostForPlatformHost("staging-platform.vellum.ai.evil.com"),
-    ).toBeNull();
   });
 });
